Ducanh

+++ src/services/category.service.ts
@@ -1,6 +1,6 @@
 import { ICategory } from '~/interfaces/category.type'
 import Category from '../models/category.model'
-
+import Product from '../models/dish.model'
 export const getAllCategoryService = async (search?: string, status?: string, page: number = 1) => {
   try {
     const query: any = {}
@@ -15,18 +15,15 @@ export const getAllCategoryService = async (search?: string, status?: string, pa
       query.status = status
     }
 
-    const limit = 6 // cố định 6 item mỗi trang
-    const skip = (page - 1) * limit
+    // const limit = 6  cố định 6 item mỗi trang
+    // const skip = (page - 1) * limit
 
-    const [data, total] = await Promise.all([
-      Category.find(query).skip(skip).limit(limit),
-      Category.countDocuments(query)
-    ])
+    const [data, total] = await Promise.all([Category.find(query), Category.countDocuments(query)])
     return {
       total,
       page,
-      limit,
-      totalPages: Math.ceil(total / limit),
+      // limit,
+      // totalPages: Math.ceil(total / limit),
       data
     }
   } catch (error) {
@@ -63,7 +60,20 @@ export const updateCategoryService = async (id: string, data: ICategory) => {
 
 export const deleteCategoryService = async (id: string) => {
   try {
-    await Category.findByIdAndDelete(id)
+    const productCount = await Product.countDocuments({ category_id: id })
+    if (productCount > 0) {
+      throw new Error('Danh mục này vẫn còn sản phẩm, không thể xóa.')
+    }
+
+    console.log('success')
+
+    const deleted = await Category.findByIdAndDelete(id)
+
+    if (!deleted) {
+      throw new Error('Không tìm thấy danh mục để xóa.')
+    }
+
+    return { message: 'Xóa danh mục thành công.' }
   } catch (error) {
     throw new Error('Cannot Delete')
   }
